pyproject.toml
[project]
name = "glifo-analise"
version = "0.2.0"
description = "Análise de viabilidade tátil para dispositivo matricial — fonte ELIS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fonttools>=4.61.1",
    "lxml>=6.0.2",
    "networkx>=3.4.2",
    "nicegui>=2.0.0",
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "trimesh>=4.11.2",
]

[project.scripts]
glifo-analise = "glifo_analise.cli.main:main"
glifo-gui     = "glifo_analise.gui.app:run"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.setuptools.packages.find]
include = ["glifo_analise*"]
